# docs/swagger.yaml
definitions:
...
  common.BaseResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  user.ProfileResponse:
    properties:
      user:
        $ref: '#/definitions/user.UserResponse'
    type: object
  user.UserResponse:
    properties:
      created_at:
        example: "2025-07-19T10:30:00Z"
        type: string
      email:
        example: john.doe@example.com
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      name:
        example: John Doe
        type: string
      role:
        example: user
        type: string
      updated_at:
        example: "2025-07-19T10:30:00Z"
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/v1/user/me:
    get:
      consumes:
      - application/json
      description: Get the current user's profile information
      produces:
      - application/json
      responses:
        "200":
          description: User profile retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/common.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/user.ProfileResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.BaseResponse'
      security:
      - ApiKeyAuth: []
      summary: Get user profile
      tags:
      - User
swagger: "2.0"